% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{poll}
\alias{poll}
\title{Add a poll to the slide}
\usage{
poll(question, options, width = "100\%", height = "500px")
}
\arguments{
\item{question}{A character -string- representing the poll question.}

\item{options}{A character vector with the possible poll answers.}

\item{width}{A character with a valid html `width` value for the iframe.}

\item{height}{A character with a valid html `height` value for the iframe.}
}
\value{
An html `iframe` tag which will show the poll results.
}
\description{
Adds a poll to the current slide. It will return the corresponding `iframe`
tag which will show the poll results. At the moment, just one poll per slide
is allowed.
}
