% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question.R
\name{question}
\alias{question}
\title{Add a question to the slide}
\usage{
question(
  question,
  allow_anonymous = TRUE,
  allow_multiple = TRUE,
  max_chars = 160,
  width = "100\%",
  height = "500px"
)
}
\arguments{
\item{question}{A character -string- representing the question.}

\item{allow_anonymous}{A logical indicating if audience can ask questions as
an anonymous user.}

\item{allow_multiple}{A logical indicating if audience can ask more than one
question.}

\item{max_chars}{A numeric indicating max characters allowed per question.
The minimum value will be 1.}

\item{width}{A character with a valid html `width` value for the iframe.}

\item{height}{A character with a valid html `height` value for the iframe.}
}
\value{
An html `iframe` tag which will show the questions asked.
}
\description{
Adds a question to the current slide. It will return the corresponding
`iframe` tag which will show the questions done.
This function must be called once per question.
}
