% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{rating}
\alias{rating}
\title{Add a rating to the slide}
\usage{
rating(question, max_value = 5, width = "100\%", height = "500px")
}
\arguments{
\item{question}{A character -string- representing the rating question.}

\item{max_value}{A numeric indicating the max value of the rate. The minimum
value will be 2.}

\item{width}{A character with a valid html `width` value for the iframe.}

\item{height}{A character with a valid html `height` value for the iframe.}
}
\value{
An html `iframe` tag which will show the rating results.
}
\description{
Adds a rating to the current slide. It will return the corresponding `iframe`
tag which will show the rating results.
This function must be called once per rating.
}
