% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_app.R
\name{set_app}
\alias{set_app}
\title{Set up the interactions server}
\usage{
set_app(app_name = "presentation", url = NULL, ...)
}
\arguments{
\item{app_name}{A character -string- with the name of the Shiny app to
create. Must be unique in your Shiny server, or it will overwrite it.}

\item{url}{A character -string- (optional) to provide a valid existing
`interactingan` server instance that was previously configured.}

\item{...}{Additional parameters passed to `rsconnect::deployApp` function.}
}
\value{
A QR code image that will point to the server url.
}
\description{
`interactingan` package creates a Shiny app to allow the interactions, this
app must be hosted on any Shiny server, by default it will try to use
[shinyapps.io](shinyapps.io).
Calling this function will reset all existing interaction objects.
}
