% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.R
\name{wordcloud}
\alias{wordcloud}
\title{Add a wordcloud to the slide}
\usage{
wordcloud(question, max_words = 2, width = "100\%", height = "500px", ...)
}
\arguments{
\item{question}{A character -string- representing the wordcloud question.}

\item{max_words}{A numeric indicating max words allowed per user.}

\item{width}{A character with a valid html `width` value for the iframe.}

\item{height}{A character with a valid html `height` value for the iframe.}

\item{...}{Additional parameters to be passed to wordcloud::wordcloud
function.}
}
\value{
An html `iframe` tag which will show the wordcloud results.
}
\description{
Adds a wordcloud to the current slide. It will return the corresponding
`iframe` tag which will show the wordcloud results.
This function must be called once per wordcloud.
}
