% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_interactions.R
\name{deploy_interactions}
\alias{deploy_interactions}
\title{Deploy all interactions to the server}
\usage{
deploy_interactions()
}
\description{
Deploys all created interaction objects to the, `interactingan` Shiny server,
previously configured by the `set_app` function.
}
