% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poll.R
\name{poll}
\alias{poll}
\title{Add a poll to the slide}
\usage{
poll(
  question,
  options,
  correct_opts = NULL,
  multiple_opts = FALSE,
  width = "100\%",
  height = "500px"
)
}
\arguments{
\item{question}{A character -string- representing the poll question.}

\item{options}{A character vector with the possible poll answers.}

\item{correct_opts}{A character vector with the correct poll answers.
Must be a subset of `options`.}

\item{multiple_opts}{A logical indicating if multiple options can be
selected.}

\item{width}{A character with a valid html `width` value for the iframe.}

\item{height}{A character with a valid html `height` value for the iframe.}
}
\value{
An html `iframe` tag which will show the poll results.
}
\description{
Adds a poll to the current slide. It will return the corresponding `iframe`
tag which will show the poll results.
This function must be called once per poll.
}
